+++ index.bare.js
@@ -1,0 +1,99 @@
+/* global Bare */
+const Signal = require('bare-signals')
+const safetyCatch = require('safety-catch')
+
+module.exports = goodbye
+
+const handlers = []
+
+let exitCode = 0
+let forceExit = false
+
+goodbye.exiting = false
+goodbye.exit = exit
+
+const sigint = new Signal('SIGINT')
+const sigterm = new Signal('SIGTERM')
+
+sigint
+  .on('signal', () => {
+    sigint.stop()
+    // TODO: uncomment
+    // sigint.unref()
+    exitCode = 130
+    onexit()
+  })
+
+sigterm
+  .on('signal', () => {
+    sigterm.stop()
+    // TODO: uncomment
+    // sigterm.unref()
+    exitCode = 130
+    onexit()
+  })
+
+function onexit () {
+  if (goodbye.exiting) return
+  goodbye.exiting = true
+
+  Bare.removeListener('exit', onexit)
+
+  const order = []
+
+  for (const h of handlers.sort((a, b) => b.position - a.position)) {
+    if (!order.length || order[order.length - 1][0].position !== h.position) order.push([])
+    order[order.length - 1].push(h)
+  }
+
+  loop()
+
+  function loop () {
+    if (!order.length) return done()
+    Promise.allSettled(order.pop().map(run)).then(loop, loop)
+  }
+
+  function done () {
+    if (forceExit) Bare.exit(exitCode)
+  }
+}
+
+async function run (h) {
+  try {
+    await h.fn()
+  } catch (e) {
+    safetyCatch(e)
+  }
+}
+
+function setup () {
+  Bare.prependListener('exit', onexit)
+  sigint.start()
+  sigterm.start()
+}
+
+function cleanup () {
+  Bare.removeListener('exit', onexit)
+  sigint.close()
+  sigterm.close()
+}
+
+function goodbye (fn, position = 0) {
+  if (handlers.length === 0) setup()
+  const handler = { position, fn }
+  handlers.push(handler)
+
+  return function unregister () {
+    const i = handlers.indexOf(handler)
+    if (i > -1) handlers.splice(i, 1)
+    if (!handlers.length) cleanup()
+  }
+}
+
+function exit () {
+  forceExit = true
+  sigint.close()
+  sigterm.close()
+  onexit()
+}
+
