+++ package.json
@@ -5,10 +5,16 @@
   "main": "index.js",
   "browser": "./browser.js",
   "dependencies": {
+    "bare-signals": "^2.1.0",
     "safety-catch": "^1.0.2"
   },
   "devDependencies": {
     "standard": "^17.0.0"
+  },
+  "imports": {
+    "bare-signals": {
+      "bare": "bare-signals"
+    }
   },
   "repository": {
     "type": "git",
